Adding ability to locally host Lato font
Under GDPR it can be preferred to locally host the Lato font
instead of including it from Google CDN.
As using the Google CDN can be beneficial performance wise, this
option stays the default.

In case a user chooses to use a font other than Lato, the option
to not include the Lato font was included as well.

Choosing the mode can be done by setting a variable.